+++ internal/plugin/algorithm.go
