+++ internal/plugin/algorithm_test.go
