+++ internal/plugin/errors_test.go
