+++ internal/plugin/manager/manager_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/notaryproject/notation-go/dir"
 	"github.com/notaryproject/notation-go/internal/mock/mockfs"
-	"github.com/notaryproject/notation-go/plugin"
+	"github.com/notaryproject/notation-go/internal/plugin"
 )
 
 type smartTestCommander struct {
@@ -46,7 +46,7 @@ var validMetadataBar = plugin.Metadata{
 }
 
 func TestManager_Get_Empty(t *testing.T) {
-	mgr := &Manager{mockfs.NewSysFSMock(fstest.MapFS{}, ""), nil}
+	mgr := &Manager{mockfs.NewSysFSMock(fstest.MapFS{}), nil}
 	got, err := mgr.Get(context.Background(), "foo")
 	if !errors.Is(err, ErrNotFound) {
 		t.Errorf("Manager.Get() error = %v, want %v", got, ErrNotFound)
@@ -69,28 +69,28 @@ func TestManager_Get_NotFound(t *testing.T) {
 	ctx := context.Background()
 
 	// empty fsys.
-	mgr := &Manager{mockfs.NewSysFSMock(fstest.MapFS{}, ""), nil}
+	mgr := &Manager{mockfs.NewSysFSMock(fstest.MapFS{}), nil}
 	check(mgr.Get(ctx, "foo"))
 
 	// plugin directory exists without executable.
 
 	mgr = &Manager{mockfs.NewSysFSMock(fstest.MapFS{
 		"foo": &fstest.MapFile{Mode: fs.ModeDir},
-	}, ""), nil}
+	}), nil}
 	check(mgr.Get(ctx, "foo"))
 
 	// plugin directory exists with symlinked executable.
 	mgr = &Manager{mockfs.NewSysFSMock(fstest.MapFS{
 		"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 		addExeSuffix("foo/notation-foo"): &fstest.MapFile{Mode: fs.ModeSymlink},
-	}, ""), nil}
+	}), nil}
 	check(mgr.Get(ctx, "foo"))
 
 	// valid plugin exists but is not the target.
 	mgr = &Manager{mockfs.NewSysFSMock(fstest.MapFS{
 		"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 		addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-	}, ""), testCommander{metadataJSON(validMetadata), true, nil}}
+	}), testCommander{metadataJSON(validMetadata), true, nil}}
 	check(mgr.Get(ctx, "baz"))
 }
 
@@ -111,7 +111,7 @@ func TestManager_Get(t *testing.T) {
 				fstest.MapFS{
 					"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 					addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-				}, ""),
+				}),
 				testCommander{nil, false, errors.New("failed")}},
 			args{"foo"},
 			&Plugin{Path: addExeSuffix("foo/notation-foo")},
@@ -123,7 +123,7 @@ func TestManager_Get(t *testing.T) {
 				fstest.MapFS{
 					"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 					addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-				}, ""),
+				}),
 				testCommander{[]byte("content"), true, nil}},
 			args{"foo"},
 			&Plugin{Path: addExeSuffix("foo/notation-foo")},
@@ -135,7 +135,7 @@ func TestManager_Get(t *testing.T) {
 				fstest.MapFS{
 					"baz":                            &fstest.MapFile{Mode: fs.ModeDir},
 					addExeSuffix("baz/notation-baz"): new(fstest.MapFile),
-				}, ""),
+				}),
 				testCommander{metadataJSON(validMetadata), true, nil}},
 			args{"baz"},
 			&Plugin{Metadata: validMetadata, Path: addExeSuffix("baz/notation-baz")},
@@ -147,7 +147,7 @@ func TestManager_Get(t *testing.T) {
 				fstest.MapFS{
 					"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 					addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-				}, ""),
+				}),
 				testCommander{metadataJSON(plugin.Metadata{Name: "foo"}), true, nil}},
 			args{"foo"},
 			&Plugin{Metadata: plugin.Metadata{Name: "foo"}, Path: addExeSuffix("foo/notation-foo")},
@@ -159,7 +159,7 @@ func TestManager_Get(t *testing.T) {
 				fstest.MapFS{
 					"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 					addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-				}, ""),
+				}),
 				testCommander{metadataJSON(validMetadata), true, nil}},
 			args{"foo"},
 			&Plugin{Metadata: validMetadata, Path: addExeSuffix("foo/notation-foo")}, "",
@@ -202,16 +202,16 @@ func TestManager_List(t *testing.T) {
 		want []*Plugin
 	}{
 		{"empty fsys",
-			&Manager{mockfs.NewSysFSMock(fstest.MapFS{}, ""), nil}, nil},
+			&Manager{mockfs.NewSysFSMock(fstest.MapFS{}), nil}, nil},
 		{"fsys without plugins",
-			&Manager{mockfs.NewSysFSMock(fstest.MapFS{"a.go": &fstest.MapFile{}}, ""), nil}, nil},
+			&Manager{mockfs.NewSysFSMock(fstest.MapFS{"a.go": &fstest.MapFile{}}), nil}, nil},
 		{"fsys with plugins but symlinked",
 			&Manager{
 				mockfs.NewSysFSMock(fstest.MapFS{
 					"foo":                            &fstest.MapFile{Mode: fs.ModeDir | fs.ModeSymlink},
 					addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
 					"baz":                            &fstest.MapFile{Mode: fs.ModeDir},
-				}, ""),
+				}),
 				testCommander{metadataJSON(validMetadata), true, nil}}, nil},
 		{
 			"fsys with some invalid plugins",
@@ -219,7 +219,7 @@ func TestManager_List(t *testing.T) {
 				mockfs.NewSysFSMock(fstest.MapFS{
 					"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 					addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-				}, ""),
+				}),
 				testCommander{metadataJSON(validMetadata), true, nil}}, []*Plugin{{Metadata: validMetadata}}},
 		{
 			"fsys with plugins",
@@ -228,7 +228,7 @@ func TestManager_List(t *testing.T) {
 					"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 					addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
 					"baz":                            &fstest.MapFile{Mode: fs.ModeDir},
-				}, ""),
+				}),
 				testCommander{metadataJSON(validMetadata), true, nil}}, []*Plugin{{Metadata: validMetadata}}},
 	}
 	for _, tt := range tests {
@@ -247,7 +247,7 @@ func TestManager_List(t *testing.T) {
 }
 
 func TestManager_Runner_Run_NotFound(t *testing.T) {
-	mgr := &Manager{mockfs.NewSysFSMock(fstest.MapFS{}, ""), nil}
+	mgr := &Manager{mockfs.NewSysFSMock(fstest.MapFS{}), nil}
 	_, err := mgr.Runner("foo")
 	if !errors.Is(err, ErrNotFound) {
 		t.Fatalf("Manager.Runner() error = %v, want %v", err, ErrNotFound)
@@ -271,7 +271,7 @@ func TestManager_Runner_Run(t *testing.T) {
 				fstest.MapFS{
 					"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 					addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-				}, ""),
+				}),
 			&testCommander{nil, false, errExec}},
 			args{"foo", plugin.CommandGenerateSignature}, errExec,
 		},
@@ -281,7 +281,7 @@ func TestManager_Runner_Run(t *testing.T) {
 					fstest.MapFS{
 						"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 						addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-					}, ""),
+					}),
 				&testCommander{[]byte("{\"errorCode\": \"ERROR\"}"), false, nil}},
 			args{"foo", plugin.CommandGenerateSignature}, plugin.RequestError{Code: plugin.ErrorCodeGeneric},
 		},
@@ -291,7 +291,7 @@ func TestManager_Runner_Run(t *testing.T) {
 					fstest.MapFS{
 						"foo":                            &fstest.MapFile{Mode: fs.ModeDir},
 						addExeSuffix("foo/notation-foo"): new(fstest.MapFile),
-					}, ""),
+					}),
 				&testCommander{metadataJSON(validMetadata), true, nil}},
 			args{"foo", plugin.CommandGenerateSignature}, nil,
 		},
