+++ internal/plugin/manager/testdata/main.go
