+++ internal/plugin/metadata.go
