+++ internal/plugin/metadata_test.go
