+++ plugin/proto/metadata_test.go
