+++ signature/plugin_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/notaryproject/notation-core-go/signature/cose"
 	"github.com/notaryproject/notation-core-go/signature/jws"
 	"github.com/notaryproject/notation-go"
-	"github.com/notaryproject/notation-go/plugin"
+	"github.com/notaryproject/notation-go/internal/plugin"
 	gcose "github.com/veraison/go-cose"
 )
